+++ src/main/java/com/upem/fr/ressource/SecteurRessource.java
@@ -34,49 +34,58 @@ public class SecteurRessource {
 
     @GetMapping("secteurs/{id}")
     public Optional<Secteur> getOne(@PathVariable Long id) {
-        try {
-            return secteurService.getOne(id);
-        }catch (NotFoundException e){
+        if(!secteurService.getOne(id).isPresent())
             throw new ResponseStatusException(HttpStatus.NOT_FOUND,"  l'id  "+  id  +  " inconnu");
-        }
+
+
+        return secteurService.getOne(id);
+
     }
 
     @DeleteMapping("secteurs/{id}")
     public void delete(@PathVariable Long id) {
-        try {
-            secteurService.delete(id);
-        }catch (NotFoundException e){
+        if(!secteurService.getOne(id).isPresent())
             throw new ResponseStatusException(HttpStatus.NOT_FOUND,"  l'id  "+  id  +  " inconnu");
-        }
+
+
+        secteurService.delete(id);
+
     }
 
     @PostMapping("secteurs/{id}")
     public Secteur update(@Valid @PathVariable Long id, @RequestBody Secteur secteur) {
-        try {
-            return secteurService.update(id, secteur);
-        }catch (NotFoundException e){
+        if(!secteurService.getOne(id).isPresent())
             throw new ResponseStatusException(HttpStatus.NOT_FOUND,"  l'id  "+  id  +  " inconnu");
-        }
+
+
+        return secteurService.update(id, secteur);
+
     }
 
     @GetMapping("secteurBassin/{secteurId}/{bassinId}")
     public Iterable<Secteur> affectBassintosecteur(@PathVariable Long secteurId, @PathVariable Long bassinId) {
-        try {
-            secteurService.addBassin(secteurService.getOne(secteurId), bassinService.getOne(bassinId));
-            return secteurService.getAll();
-        }catch (NotFoundException e){
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"SecteurId "+  secteurId  +
-                    " inconnu or BassinId"+bassinId+"inconnu");
-        }
+        if(!secteurService.getOne(secteurId).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"  l'id  "+  secteurId  +  " inconnu");
+
+        if(!bassinService.getOne(bassinId).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"  l'id  "+  bassinId  +  " inconnu");
+
+
+        secteurService.addBassin(secteurService.getOne(secteurId), bassinService.getOne(bassinId));
+        return secteurService.getAll();
+
     }
     @GetMapping("deleteBassinSecteur/{secteurId}/{bassinId}")
     public Iterable<Secteur> deleteBassinInSecteur(@PathVariable Long secteurId, @PathVariable Long bassinId) {
-        try {
-            secteurService.removeBassin(secteurService.getOne(secteurId), bassinService.getOne(bassinId));
-            return secteurService.getAll();
-        }catch (NotFoundException e){
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"SecteurId "+  secteurId  +
-                    " inconnu or BassinId"+bassinId+"inconnu");
-        }
+        if(!secteurService.getOne(secteurId).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"  l'id  "+  secteurId  +  " inconnu");
+
+        if(!bassinService.getOne(bassinId).isPresent())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"  l'id  "+  bassinId  +  " inconnu");
+
+
+        secteurService.removeBassin(secteurService.getOne(secteurId), bassinService.getOne(bassinId));
+        return secteurService.getAll();
+
     }
 }
